tests: increase unit test coverage and move Types to separate file
